+++ libs/web-ui/src/index.js
@@ -1,0 +1,1 @@
+export  {TopNavbar}  from './components/TopNavbar'